Use qp_num to identify sender
ibv_wc::qp_num is used to identify the sender of a message. A map of
qp_num -> rdma_cm_id * is maintained. Upon receipt of a message, the
qp_num of the ibv_wc is used to look up the rdma_cm_id * of the sender.
The qp_num is returned to the caller inside a future, returned by
async_rdma_operation. For rdma_post_read() operations the returned
future has no value other than indicate the end of the operation (which
was it's only use anyway). This could be changed, by return a
future<pair<qp_t, T>>, so that qp_num as well as a pointer to the used
buffer is returned.